requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==2.1.0
requests==2.32.3
pytz==2024.1
python-dateutil==2.9.0.post0
nsepython==1.6.3
nsetools==1.0.11
yfinance==0.2.43
nselib==0.1.0
beautifulsoup4==4.12.3
streamlit-autorefresh==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0

